UPSTREAM: cbmem: add coreboot table records for each cbmem entry
In order to not expose the cbmem data structures to userland
that are used by coreboot internally add each of the cbmem
entries to a coreboot table record. The payload ABI uses
coreboot tables so this just provides a shortcut for cbmem
entries which were manually added previously by doing the
work on behalf of all entries.

A cursor structure and associated functions are added to
the imd code for walking the entries in order to be placed
in the coreboot tables.  Additionally a struct lb_cbmem_entry
is added that lists the base address, size, and id of the
cbmem entry.

BUG=chrome-os-partner:43731
BRANCH=None
TEST=Booted glados. View coreboot table entries with cbmem.

Original-Change-Id: I125940aa1898c3e99077ead0660eff8aa905b13b
Original-Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-on: http://review.coreboot.org/11757
Original-Reviewed-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>
Original-Tested-by: build bot (Jenkins)
(cherry picked from commit 1ca2d864dd09788ab69f461074889bab57a92ae8)

Change-Id: I0a8b978e7f8b7daeeccab1d93571590a52a2ef24
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/310571
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>